Clean-up the code and split some code for QoL
