+++ app.js
@@ -17,7 +17,7 @@ const options = {
 }
 
 const { connect } = require("./db/connection")
-const { apiRouter } = require("./routes/api")
+const { apiRouter } = require("./routes/bookmarks")
 const { authRouter } = require("./routes/auth")
 const { docsRouter } = require("./routes/docs")
 const { usersRouter } = require("./routes/users")
