+++ helpers/checkPassword.js
@@ -1,0 +1,7 @@
+const bcrypt = require("bcrypt")
+
+const checkPassword = async (plainTextPassword, hashedPassword) => {
+  return await bcrypt.compare(plainTextPassword, hashedPassword)
+}
+
+module.exports = checkPassword
