+++ middleware/validation.js
@@ -1,4 +1,6 @@
 const { check } = require("express-validator/check")
+
+// Helpers
 const createError = require("../helpers/createError")
 const { emptyBody } = require("../helpers/errorMessages")
